agentic-reviewer/requirements.txt
pydantic>=2.0
pyyaml>=6.0
fastapi>=0.100
uvicorn>=0.22
httpx>=0.24
python-dotenv>=1.0
